requirements.txt
ollama
weaviate-client